+++ app/routes/api/portfolio/$slug.ts
@@ -1,4 +1,5 @@
-import { json, LoaderFunction } from "remix";
+import { json } from "remix";
+import type { LoaderFunction } from "remix";
 
 import { fetchFromGraphCMS, gql } from "~/utils/graphcms";
